WIP genericarm64: For debug, support booting w/o initrd on two YP ref boards.
DOES NOT WORK on BeaglePlay!!

The TI based BeaglePlay and the AMD Xilinx kv260 are the selected
reference boards for Yocto Project and genericarm64.

For these two boards only, support booting without the initrd to help with
debugging initrd issues.

In general genericarm64 supports laodable modules (=m) for everything that
can be instead of built in (=y).  However, here we make an exception for
serial port and MMC drivers for the above two platforms only.

Even for these platforms, the normal flow is to use initrd with the
genericarm64 stock kernel.

Signed-off-by: Bill Mills <bill.mills@linaro.org>